+++ packages/scripts/package.json
@@ -11,7 +11,6 @@
     "sampleAndSaveMarkdownPages": "npm run build && node ./build/sampleAndSaveMarkdownPages.js",
     "analyzeMessages": "npm run build && node ./build/analyzeMessages.js",
     "findFaq": "npm run build && node ./build/main/findFaqMain.js",
-    "importVerifiedAnswers": "npm run build && node ./build/main/importVerifiedAnswersMain.js",
     "createView/messages_by_rating": "npm run build && node ./build/createView/messages_by_rating.js",
     "createView/top_250_references": "npm run build && node ./build/createView/top_250_references.js",
     "createView/scrubbed_messages_by_rating": "npm run build && node ./build/createView/scrubbed_messages_by_rating.js",
